+++ examples/Console/TestJaegerExporter.cs
@@ -13,6 +13,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 // </copyright>
+using System.Diagnostics;
 using OpenTelemetry;
 using OpenTelemetry.Resources;
 using OpenTelemetry.Trace;
@@ -23,11 +24,29 @@ namespace Examples.Console
     {
         internal static object Run(string host, int port)
         {
+            // Prerequisite for running this example.
+            // Setup Jaegar inside local docker using following command (Source: https://www.jaegertracing.io/docs/1.21/getting-started/#all-in-one):
+            /*
+            $ docker run -d --name jaeger \
+            -e COLLECTOR_ZIPKIN_HTTP_PORT=9411 \
+            -p 5775:5775/udp \
+            -p 6831:6831/udp \
+            -p 6832:6832/udp \
+            -p 5778:5778 \
+            -p 16686:16686 \
+            -p 14268:14268 \
+            -p 14250:14250 \
+            -p 9411:9411 \
+            jaegertracing/all-in-one:1.21
+            */
+
             // To run this example, run the following command from
             // the reporoot\examples\Console\.
             // (eg: C:\repos\opentelemetry-dotnet\examples\Console\)
             //
             // dotnet run jaeger -h localhost -p 6831
+            // For non-Windows (e.g., MacOS)
+            // dotnet run jaeger -- -h localhost -p 6831
             return RunWithActivity(host, port);
         }
 
@@ -42,6 +61,21 @@ namespace Examples.Console
                     {
                         o.AgentHost = host;
                         o.AgentPort = port;
+
+                        // Examples for the rest of the options, defaults unless otherwise specified
+                        // Omitting Process Tags example as Resource API is recommended for additional tags
+                        o.MaxPayloadSizeInBytes = 4096;
+
+                        // Using Batch Exporter (which is default)
+                        // The other option is ExportProcessorType.Simple
+                        o.ExportProcessorType = ExportProcessorType.Batch;
+                        o.BatchExportProcessorOptions = new BatchExportProcessorOptions<Activity>()
+                        {
+                            MaxQueueSize = 2048,
+                            ScheduledDelayMilliseconds = 5000,
+                            ExporterTimeoutMilliseconds = 30000,
+                            MaxExportBatchSize = 512,
+                        };
                     })
                     .Build();