+++ src/OpenTelemetry.Exporter.InMemory/CHANGELOG.md
@@ -1,6 +1,29 @@
 # Changelog
 
+## Experimental - Metrics
+
+Note: the metrics work is happening in the [metrics feature
+branch](https://github.com/open-telemetry/opentelemetry-dotnet/tree/metrics),
+please check the latest changes
+[here](https://github.com/open-telemetry/opentelemetry-dotnet/blob/metrics/src/OpenTelemetry.Exporter.InMemory/CHANGELOG.md#experimental---metrics).
+
 ## Unreleased
+
+## 1.0.1
+
+Released 2021-Feb-10
+
+## 1.0.0-rc4
+
+Released 2021-Feb-09
+
+## 1.0.0-rc3
+
+Released 2021-Feb-04
+
+## 1.0.0-rc2
+
+Released 2021-Jan-29
 
 * `AddInMemoryExporter` extension method for traces moved from
   `OpenTelemetry` namespace to `OpenTelemetry.Trace` namespace.